+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Add dependabot update configuration
+
 ## [0.6.0] - 2024-10-06
 
 ### Added