+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.7.0] - 2025-01-02
+
 ### Added
 
 - ~~Add dependabot update configuration~~